requirements.txt
mysql-connector-python==8.0.28
pandas==2.2.2
numpy==1.26
python-dotenv==1.0.1
requests==2.32.3
SQLAlchemy==2.0.34
